# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DSA Demo API",
    "version": "1.0.0",
    "description": "API for retrieving swim school data, including branches, users, students, levels, skills, classes, sessions, enrollments, assessments, and attendance."
  },
  "servers": [
    { "url": "https://dsa-demo.onrender.com" }
  ],
  "paths": {
    "/branches": {
      "get": {
        "operationId": "listBranches",
        "summary": "Get all branches",
        "tags": ["Branches"],
        "responses": {
          "200": { "description": "List of branches" }
        }
      }
    },
    "/branches/{id}": {
      "get": {
        "operationId": "getBranchById",
        "summary": "Get branch by ID",
        "tags": ["Branches"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Branch details" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/users": {
      "get": {
        "operationId": "listUsers",
        "summary": "Get all users (staff)",
        "tags": ["Users"],
        "parameters": [
          { "in": "query", "name": "role", "schema": { "type": "string" } },
          { "in": "query", "name": "branchId", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of users" }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "getUserById",
        "summary": "Get user by ID",
        "tags": ["Users"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "User details" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/students": {
      "get": {
        "operationId": "listStudents",
        "summary": "Get all students",
        "tags": ["Students"],
        "parameters": [
          { "in": "query", "name": "levelId", "schema": { "type": "string" } },
          { "in": "query", "name": "branchId", "schema": { "type": "string" } },
          { "in": "query", "name": "status", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of students" }
        }
      }
    },
    "/students/{id}": {
      "get": {
        "operationId": "getStudentById",
        "summary": "Get student by ID",
        "tags": ["Students"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Student details" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/students/{id}/enrollments": {
      "get": {
        "operationId": "listStudentEnrollments",
        "summary": "Get all enrollments for a student",
        "tags": ["Students", "Enrollments"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "List of enrollments" }
        }
      }
    },
    "/students/{id}/assessments": {
      "get": {
        "operationId": "listStudentAssessments",
        "summary": "Get all assessments for a student",
        "tags": ["Students", "Assessments"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "List of assessments" }
        }
      }
    },
    "/students/{id}/attendances": {
      "get": {
        "operationId": "listStudentAttendances",
        "summary": "Get all attendance records for a student",
        "tags": ["Students", "Attendance"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "List of attendance records" }
        }
      }
    },
    "/levels": {
      "get": {
        "operationId": "listLevels",
        "summary": "Get all levels",
        "tags": ["Levels"],
        "responses": {
          "200": { "description": "List of levels" }
        }
      }
    },
    "/levels/{id}": {
      "get": {
        "operationId": "getLevelById",
        "summary": "Get level by ID",
        "tags": ["Levels"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Level details" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/levels/{id}/skills": {
      "get": {
        "operationId": "listLevelSkills",
        "summary": "Get all skills for a level",
        "tags": ["Levels", "Skills"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "List of skills" }
        }
      }
    },
    "/skills": {
      "get": {
        "operationId": "listSkills",
        "summary": "Get all skills",
        "tags": ["Skills"],
        "parameters": [
          { "in": "query", "name": "levelId", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of skills" }
        }
      }
    },
    "/class-types": {
      "get": {
        "operationId": "listClassTypes",
        "summary": "Get all class types",
        "tags": ["Class Types"],
        "responses": {
          "200": { "description": "List of class types" }
        }
      }
    },
    "/class-types/{id}": {
      "get": {
        "operationId": "getClassTypeById",
        "summary": "Get class type by ID",
        "tags": ["Class Types"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Class type details" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/swim-classes": {
      "get": {
        "operationId": "listSwimClasses",
        "summary": "Get all swim classes",
        "tags": ["Swim Classes"],
        "parameters": [
          { "in": "query", "name": "branchId", "schema": { "type": "string" } },
          { "in": "query", "name": "classTypeId", "schema": { "type": "string" } },
          { "in": "query", "name": "levelId", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of swim classes" }
        }
      }
    },
    "/swim-classes/{id}": {
      "get": {
        "operationId": "getSwimClassById",
        "summary": "Get swim class by ID",
        "tags": ["Swim Classes"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Swim class details" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/swim-classes/{id}/sessions": {
      "get": {
        "operationId": "listSwimClassSessions",
        "summary": "Get all sessions for a swim class",
        "tags": ["Swim Classes", "Sessions"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "List of sessions" }
        }
      }
    },
    "/swim-class-sessions": {
      "get": {
        "operationId": "listSwimClassSessionsAll",
        "summary": "Get all swim class sessions",
        "tags": ["Sessions"],
        "parameters": [
          { "in": "query", "name": "swimClassId", "schema": { "type": "string" } },
          { "in": "query", "name": "dateFrom", "schema": { "type": "string", "format": "date" } },
          { "in": "query", "name": "dateTo", "schema": { "type": "string", "format": "date" } }
        ],
        "responses": {
          "200": { "description": "List of swim class sessions" }
        }
      }
    },
    "/swim-class-sessions/{id}": {
      "get": {
        "operationId": "getSwimClassSessionById",
        "summary": "Get swim class session by ID",
        "tags": ["Sessions"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Swim class session details" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/swim-class-sessions/{id}/attendances": {
      "get": {
        "operationId": "listSessionAttendances",
        "summary": "Get all attendance records for a session",
        "tags": ["Sessions", "Attendance"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "List of attendance records" }
        }
      }
    },
    "/enrollments": {
      "get": {
        "operationId": "listEnrollments",
        "summary": "Get all enrollments",
        "tags": ["Enrollments"],
        "parameters": [
          { "in": "query", "name": "studentId", "schema": { "type": "string" } },
          { "in": "query", "name": "classId", "schema": { "type": "string" } },
          { "in": "query", "name": "active", "schema": { "type": "boolean" } }
        ],
        "responses": {
          "200": { "description": "List of enrollments" }
        }
      }
    },
    "/enrollments/{id}": {
      "get": {
        "operationId": "getEnrollmentById",
        "summary": "Get enrollment by ID",
        "tags": ["Enrollments"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Enrollment details" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/assessments": {
      "get": {
        "operationId": "listAssessments",
        "summary": "Get all assessments",
        "tags": ["Assessments"],
        "parameters": [
          { "in": "query", "name": "studentId", "schema": { "type": "string" } },
          { "in": "query", "name": "levelId", "schema": { "type": "string" } },
          { "in": "query", "name": "status", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of assessments" }
        }
      }
    },
    "/assessments/{id}": {
      "get": {
        "operationId": "getAssessmentById",
        "summary": "Get assessment by ID",
        "tags": ["Assessments"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Assessment details" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/attendances": {
      "get": {
        "operationId": "listAttendances",
        "summary": "Get all attendance records",
        "tags": ["Attendance"],
        "parameters": [
          { "in": "query", "name": "studentId", "schema": { "type": "string" } },
          { "in": "query", "name": "swimClassSessionId", "schema": { "type": "string" } },
          { "in": "query", "name": "confirmed", "schema": { "type": "boolean" } }
        ],
        "responses": {
          "200": { "description": "List of attendance records" }
        }
      }
    },
    "/attendances/{id}": {
      "get": {
        "operationId": "getAttendanceById",
        "summary": "Get attendance by ID",
        "tags": ["Attendance"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Attendance details" },
          "404": { "description": "Not found" }
        }
      }
    }
  }
}